+++ lib/src/wiggle_abi.rs
@@ -50,6 +50,8 @@ macro_rules! multi_value_result {
 
 mod backend_impl;
 mod body_impl;
+mod cache;
+mod config_store;
 mod dictionary_impl;
 mod entity;
 mod fastly_purge_impl;
@@ -70,9 +72,9 @@ wiggle::from_witx!({
     errors: { fastly_status => Error },
     async: {
         fastly_async_io::{select},
-        fastly_object_store::{insert},
+        fastly_object_store::{insert, lookup_async, pending_lookup_wait},
         fastly_http_body::{append, read, write},
-        fastly_http_req::{pending_req_select, pending_req_wait, send},
+        fastly_http_req::{pending_req_select, pending_req_poll, pending_req_wait, send, send_async, send_async_streaming},
     }
 });