+++ test-fixtures/src/bin/async_io.rs
@@ -2,11 +2,13 @@
 //! a distinct backend. Checks each for readiness, and then does a select with timeout against
 //! all of them.
 
+use std::io::Write;
 use std::str::FromStr;
 
 use fastly::handle::{BodyHandle, RequestHandle, ResponseHandle};
 use fastly::http::{HeaderName, HeaderValue, Method, StatusCode, Url};
 use fastly::Error;
+use fastly::Request;
 use fastly_shared::{CacheOverride, FastlyStatus};
 
 fn is_ready(handle: u32) -> bool {
@@ -44,6 +46,12 @@ fn test_select() -> Result<(), Error> {
     let (_read_body_resp, read_body) = read_body_req.send(BodyHandle::new(), "ReadBody")?;
     let read_body_handle = unsafe { read_body.as_u32() };
 
+    // This request is used as a synchronization mechanism for the purposes of
+    // this test
+    let write_body_sync_req = Request::get("http://writebody.org/")
+        .send_async("Semaphore")
+        .expect("request begins sending");
+
     // The "write body" case involves a `send_async_streaming` call, where the streaming body is the
     // async item of interest. To test readiness, we need to ensure the body is large enough that Hyper
     // won't try to buffer it, and hence we can see backpressure on streaming. We do this by including
@@ -54,7 +62,9 @@ fn test_select() -> Result<(), Error> {
     write_body_req.set_cache_override(&pass);
     write_body_req.set_method(&Method::POST);
     let mut write_body_initial = BodyHandle::new();
-    let initial_bytes = write_body_initial.write_bytes(&vec![0; INITIAL_BYTE_COUNT]);
+    let initial_bytes = write_body_initial
+        .write(&vec![0; INITIAL_BYTE_COUNT])
+        .expect("failed to write to body handle");
     assert_eq!(initial_bytes, INITIAL_BYTE_COUNT);
     let (mut write_body, _write_body_pending_req) =
         write_body_req.send_async_streaming(write_body_initial, "WriteBody")?;
@@ -65,17 +75,18 @@ fn test_select() -> Result<(), Error> {
     // body we are streaming to it.
     let one_chunk = vec![0; 8 * 1024];
     while is_ready(write_body_handle) {
-        let nwritten = write_body.write_bytes(&one_chunk);
+        let nwritten = write_body
+            .write(&one_chunk)
+            .expect("failed to write to streaming body handle");
         assert!(nwritten > 0);
     }
 
-    // Give the servers a chance to do their thing. This is needed to resolve a race between the servers
-    // initiating responses / reading buffers and the guest snapshotting readiness or performing `select`
-    //
-    // If this ever becomes flaky, an alternative would be to introduce an additional backend that
-    // responds only when the other backends have reached their steady state, via a VTC semaphore, and
-    // block on THAT backend here.
-    std::thread::sleep(std::time::Duration::from_millis(100));
+    // We wait on this request here to give the servers a chance to do their
+    // thing. This is needed to resolve a race between the servers initiating
+    // responses / reading buffers and the guest snapshotting readiness or
+    // performing `select`. This request should return when the other backends
+    // have reached their steady state
+    write_body_sync_req.wait()?;
 
     append_header(
         &mut ds_resp,